[INJIVER-139] Implemented Stepper component Signed-off-by: Shiva <shivakumar@beehyv.com>
