--- inji-verify/src/components/Home/VerificationProgressTracker/InjiStepper/StepperConnector.tsx
@@ -1,5 +1,0 @@
-import styled from "@emotion/styled";
-
-export const StepperConnector = styled('div')`
-    border-color: #FF8F00;
-`
