--- inji-verify/src/components/Home/VerificationProgressTracker/InjiStepper/StepperIcon.tsx
@@ -1,13 +1,0 @@
-import {StepperIconContainer} from "./styles";
-
-const StepperIcon = (props: any) => {
-    const { active, completed, className } = props;
-
-    return (
-        <StepperIconContainer ownerState={{ completed, active }} className={className}>
-            {props.icon}
-        </StepperIconContainer>
-    );
-}
-
-export default StepperIcon;
