+++ inji-verify/tailwind.config.js
@@ -5,6 +5,9 @@ module.exports = {
   ],
   theme: {
     extend: {},
+    fontFamily: {
+      "inter": ["Inter"]
+    }
   },
   plugins: [],
 }
